[ADF-3177] renamed folder and style changes
