+++ lib/core/layout/components/header/header.component.html
@@ -3,7 +3,7 @@
         <mat-icon class="mat-icon material-icons" role="img" aria-hidden="true">menu</mat-icon>
     </button>
 
-    <img *ngIf="logo" src="{{logo}}" class="adf-app-logo">
+    <img src="{{logo}}" class="adf-app-logo">
 
     <span fxFlex="1 1 auto" fxShow fxHide.lt-sm="true" class="adf-app-title">{{title}}</span>
     <ng-content></ng-content>
