+++ lib/core/layout/components/header/header.component.scss
@@ -8,17 +8,25 @@
         color: mat-color($primary, default-contrast) !important;
         padding: 0 24px;
 
+        button {
+            width: 24px;
+            margin-right: 16px;
+        }
+
         .adf-app-logo {
             height: 28px;
             width: 28px;
-            padding-right: 8px;
-            padding-left: 16px;
+            margin-right: 8px;
         }
 
         .adf-header-delimiter {
             height: 24px;
             width: 2px;
             background-color: mat-color($primary, default-contrast);
+            margin-right: 16px;
+        }
+
+        .adf-app-layout-user-profile {
             margin-right: 16px;
         }
