+++ lib/core/layout/components/layout-container/layout-container.component.scss
