+++ lib/core/layout/components/sidebar-action/sidebar-action-menu.component.scss
