+++ lib/core/layout/components/sidenav-layout/sidenav-layout.component.scss
