+++ lib/core/layout/directives/sidenav-layout-content.directive.ts
