+++ lib/core/layout/directives/sidenav-layout-header.directive.ts
