+++ lib/core/layout/directives/sidenav-layout-navigation.directive.ts
