+++ lib/core/styles/_index.scss
@@ -22,14 +22,14 @@
 @import '../viewer/components/txtViewer.component';
 @import '../viewer/components/imgViewer.component';
 @import '../form/components/form.component';
-@import '../sidenav-layout/components/sidebar-action/sidebar-action-menu.component';
+@import '../layout/components/sidebar-action/sidebar-action-menu.component';
 @import '../comments/comment-list.component';
 @import '../comments/comments.component';
-@import '../sidenav-layout/components/layout-container/layout-container.component';
+@import '../layout/components/layout-container/layout-container.component';
 @import "../templates/empty-content/empty-content.component";
 @import "../templates/error-content/error-content.component";
 @import "../buttons-menu/buttons-menu.component";
-@import '../sidenav-layout/components/header/header.component';
+@import '../layout/components/header/header.component';
 
 @mixin adf-core-theme($theme) {
     @include adf-colors-theme($theme);
